Common Move of isNotSameId Functions
- [ ] Common Move of isNotSameId Functions
- [ ] Use similar logic corresponding common functions
